Move SaveFromControls from ArgsTable to Controls addon
